endpoint: Set accept_local=1 and send_redirects=0 on veth
NodePort loadbalancers will cause packets to be emitted to the veth with
the source address set to the host's address. This would trigger the
martian source check in the Linux routing layer. Disable the check when
we are coming from Cilium managed veths.

Backports: f917bd19871190fcd910d6539101fad2841f1a9c
Signed-off-by: Thomas Graf <thomas@cilium.io>